fixed global/personnal path to local

+++ python_tools/fix_track_paths/fix_track_paths.sh
@@ -9,6 +9,7 @@ cp ${HOME}/.mixxx/mixxxdb.sqlite ${HOME}/.mixxx/mixxxdb.sqlite.bak.$(date +%y%m%
 
 # using the information of the custom database to modify Mixxx's database
 sqlite3  ${HOME}/.mixxx/mixxxdb.sqlite < mixxxdb_fix_tracks_locations.sql
+test $? -ne 0 && exit
 
 # deleting the custom database
 rm -f custom_music_db.sqlite
